Register, Login, Admin and Salesman menus, Sale

+++ CarDealership/Data/Models/Salesman.cs
@@ -11,6 +11,6 @@ namespace CarDealership.Data.Models
         
         public string Name { get; set; }
         public string Password { get; set; }
-        public double Profits { get; set; }
+        public decimal Profits { get; set; }
     }
 }
